+++ actix-web/CHANGES.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 4.11.0
+
 - Add `Logger::log_level()` method.
 - Improve handling of non-UTF-8 header values in `Logger` middleware.
 - Add `HttpServer::shutdown_signal()` method.